If search yields empty result status is 200, returns empty arr, query is fixed for transactions overview
