+++ app/src/Controller/CategoryController.php
@@ -26,7 +26,6 @@ class CategoryController extends AbstractController
      * Find categories by income or expenses using query params
      * - Example url: localhost:8080/api/categories?page=1&type=expense&limit=15
      */
-    #[Route(name: 'api_find_categories_by_type', methods: ['GET'])]
     #[OA\Get(
         description: 'Returns list of categories based on chosen type',
         summary: 'Find categories by income or expenses using query params',
@@ -64,6 +63,7 @@ class CategoryController extends AbstractController
         ]
     )]
     #[Security(name: 'Bearer')]
+    #[Route(name: 'api_find_categories_by_type', methods: ['GET'])]
     public function search(
         #[MapQueryString] ?CategoryQueryDto $categoryQueryDto,
         CategoryService                     $categoryService
@@ -71,14 +71,6 @@ class CategoryController extends AbstractController
     {
         //Search
         $data = $categoryService->search($categoryQueryDto);
-
-        // If no categories are found
-        if (empty($data['categories'])) {
-            return $this->json([
-                'success' => false,
-                'message' => 'No categories found'
-            ], 404);
-        }
 
         return $this->json([
             'success' => true,
@@ -89,7 +81,6 @@ class CategoryController extends AbstractController
     /**
      * Create a new category for logged-in user
      */
-    #[Route(name: 'api_add_category', methods: ['POST'])]
     #[OA\Post(
         description: 'Creates a new category with logged-in user as its owner',
         summary: 'Create a new category for logged-in user',
@@ -131,6 +122,7 @@ class CategoryController extends AbstractController
         ]
     )]
     #[Security(name: 'Bearer')]
+    #[Route(name: 'api_add_category', methods: ['POST'])]
     public function create(
         #[MapRequestPayload] CategoryCreateDto $categoryCreateDto,
         CategoryService                        $categoryService
@@ -148,7 +140,6 @@ class CategoryController extends AbstractController
         );
     }
 
-    #[Route(name: 'api_update_category', methods: ['PATCH'])]
     #[OA\Patch(
         description: 'Makes changes to existing category',
         summary: 'Update existing category',
@@ -190,6 +181,7 @@ class CategoryController extends AbstractController
             )
         ]
     )]
+    #[Route(name: 'api_update_category', methods: ['PATCH'])]
     #[Security(name: 'Bearer')]
     public function update(
         #[MapRequestPayload] CategoryUpdateDto $categoryUpdateDto,
@@ -220,7 +212,6 @@ class CategoryController extends AbstractController
      * Delete category if it belongs to user
      * - Example url: localhost:8080/api/categories/123
      */
-    #[Route('/{id<\d+>}', name: 'api_delete_category', methods: ['DELETE'])]
     #[OA\Delete(
         description: "Removes category that is in ownership of logged user",
         summary: "Deletes specific category based on given id",
@@ -254,6 +245,7 @@ class CategoryController extends AbstractController
         ]
     )]
     #[Security(name: 'Bearer')]
+    #[Route('/{id<\d+>}', name: 'api_delete_category', methods: ['DELETE'])]
     public function delete(int $id, CategoryService $categoryService): JsonResponse
     {
         $categoryService->delete($id);
